Rename file.ts to input.ts
