+++ lib/input.ts
